+++ linked_list.h
@@ -17,7 +17,7 @@ typedef struct linked_list_t {
 linked_list_t* ll_create();
 bool ll_destroy(linked_list_t *list);
 bool ll_append(linked_list_t *list, void *data);
-void* ll_get_nth_element(const linked_list_t* list, int elem);
+void* ll_get_nth_element(const linked_list_t* list, size_t elem);
 linked_list_t *string_to_list(const char *str, const char *delim, size_t max_len);
 
 #endif // LINKED_LIST_H
